Fin d'implémentation de la classe disparition

+++ src/main/java/bandeau/Rotation.java
@@ -13,7 +13,7 @@ public class Rotation extends Animations {
     
     private double angl;
     
-    public Rotation(Float angle){
+    public Rotation(Double angle){
         angl=angle;
     }
